@@ -1,13 +1,12 @@
 package tamaized.melongolem;
 
-
 import com.google.common.collect.ImmutableSet;
-import net.minecraft.block.Block;
-import net.minecraft.block.Blocks;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.Hand;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -23,9 +22,9 @@ public class ModEventListener {
 
 	@SubscribeEvent
 	public static void onRightClick(PlayerInteractEvent.RightClickBlock e) {
-		PlayerEntity player = e.getPlayer();
-		World world = e.getWorld();
-		BlockPos pos = e.getPos();
+		Player player = e.getPlayer();
+		Level world = e.getWorld();
+		BlockPos vertex = e.getPos();
 		if (MELONS == null)
 			MELONS = ImmutableSet.of(
 
@@ -35,18 +34,18 @@ public class ModEventListener {
 
 			);
 		for (Block melonCheck : MELONS) {
-			if (!world.isRemote && world.getBlockState(pos).getBlock() == melonCheck && MelonConfig.compareStabbyItem(player.getHeldItem(Hand.MAIN_HAND)) && MelonConfig.compareStabbyItem(player.getHeldItem(Hand.OFF_HAND))) {
-				if (world.getBlockState(pos.down()).getBlock() == melonCheck && world.getBlockState(pos.up()).getBlock() == melonCheck) {
+			if (!world.isClientSide && world.getBlockState(vertex).getBlock() == melonCheck && MelonConfig.compareStabbyItem(player.getItemInHand(InteractionHand.MAIN_HAND)) && MelonConfig.compareStabbyItem(player.getItemInHand(InteractionHand.OFF_HAND))) {
+				if (world.getBlockState(vertex.below()).getBlock() == melonCheck && world.getBlockState(vertex.above()).getBlock() == melonCheck) {
 					if (!player.isCreative()) {
-						player.getHeldItem(Hand.MAIN_HAND).shrink(1);
-						player.getHeldItem(Hand.OFF_HAND).shrink(1);
+						player.getItemInHand(InteractionHand.MAIN_HAND).shrink(1);
+						player.getItemInHand(InteractionHand.OFF_HAND).shrink(1);
 					}
-					world.removeBlock(pos.down(), false);
-					world.removeBlock(pos, false);
-					world.removeBlock(pos.up(), false);
+					world.removeBlock(vertex.below(), false);
+					world.removeBlock(vertex, false);
+					world.removeBlock(vertex.above(), false);
 					EntityMelonGolem melon = melonCheck == MelonMod.glisteringMelonBlock ? new EntityGlisteringMelonGolem(world) : new EntityMelonGolem(world);
-					melon.setPositionAndUpdate(pos.getX() + 0.5F, pos.getY() - 0.5F, pos.getZ() + 0.5F);
-					world.addEntity(melon);
+					melon.teleportTo(vertex.getX() + 0.5F, vertex.getY() - 0.5F, vertex.getZ() + 0.5F);
+					world.addFreshEntity(melon);
 					break;
 				}
 			}
