@@ -1,6 +1,6 @@
 package tamaized.melongolem.common;
 
-import net.minecraft.world.World;
+import net.minecraft.world.level.Level;
 import tamaized.melongolem.MelonMod;
 
 import java.util.Objects;
@@ -9,7 +9,7 @@ public class EntityGlisteringMelonGolem extends EntityMelonGolem {
 
 	//	private static final ResourceLocation LOOT = LootTables.register(new ResourceLocation(MelonMod.MODID, "glisteringmelongolem"));
 
-	public EntityGlisteringMelonGolem(World worldIn) {
+	public EntityGlisteringMelonGolem(Level worldIn) {
 		super(Objects.requireNonNull(MelonMod.entityTypeGlisteringMelonGolem), worldIn);
 	}
 
