@@ -2,57 +2,57 @@ package tamaized.melongolem.common;
 
 import com.google.common.collect.Lists;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.command.CommandException;
-import net.minecraft.entity.EntitySize;
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.IRangedAttackMob;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.MobEntity;
-import net.minecraft.entity.Pose;
-import net.minecraft.entity.ai.attributes.AttributeModifierMap;
-import net.minecraft.entity.ai.attributes.Attributes;
-import net.minecraft.entity.ai.goal.Goal;
-import net.minecraft.entity.ai.goal.LookAtGoal;
-import net.minecraft.entity.ai.goal.LookRandomlyGoal;
-import net.minecraft.entity.ai.goal.NearestAttackableTargetGoal;
-import net.minecraft.entity.ai.goal.RangedAttackGoal;
-import net.minecraft.entity.ai.goal.WaterAvoidingRandomWalkingGoal;
-import net.minecraft.entity.item.ItemEntity;
-import net.minecraft.entity.monster.IMob;
-import net.minecraft.entity.passive.GolemEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.item.ShearsItem;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.network.PacketBuffer;
-import net.minecraft.network.datasync.DataParameter;
-import net.minecraft.network.datasync.DataSerializers;
-import net.minecraft.network.datasync.EntityDataManager;
-import net.minecraft.tileentity.SignTileEntity;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.DamageSource;
-import net.minecraft.util.Direction;
-import net.minecraft.util.Hand;
-import net.minecraft.util.SoundEvent;
-import net.minecraft.util.SoundEvents;
-import net.minecraft.util.math.AxisAlignedBB;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.TextComponentUtils;
-import net.minecraft.world.World;
+import net.minecraft.commands.CommandRuntimeException;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.ComponentUtils;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.EntityDataSerializers;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.sounds.SoundEvent;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.util.Mth;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.EntityDimensions;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.Pose;
+import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.goal.Goal;
+import net.minecraft.world.entity.ai.goal.LookAtPlayerGoal;
+import net.minecraft.world.entity.ai.goal.RandomLookAroundGoal;
+import net.minecraft.world.entity.ai.goal.RangedAttackGoal;
+import net.minecraft.world.entity.ai.goal.WaterAvoidingRandomStrollGoal;
+import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.entity.monster.Enemy;
+import net.minecraft.world.entity.monster.RangedAttackMob;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ShearsItem;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.animal.AbstractGolem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.SignBlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.AABB;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.common.IForgeShearable;
 import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.fml.common.registry.IEntityAdditionalSpawnData;
-import net.minecraftforge.fml.network.PacketDistributor;
+import net.minecraftforge.fmllegacy.common.registry.IEntityAdditionalSpawnData;
+import net.minecraftforge.fmllegacy.network.PacketDistributor;
 import net.minecraftforge.items.CapabilityItemHandler;
 import net.minecraftforge.items.IItemHandler;
 import tamaized.melongolem.IModProxy;
@@ -65,24 +65,24 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.Objects;
 
-public class EntityMelonGolem extends GolemEntity implements IRangedAttackMob, IForgeShearable, IEntityAdditionalSpawnData, IModProxy.ISignHolder {
+public class EntityMelonGolem extends AbstractGolem implements RangedAttackMob, IForgeShearable, IEntityAdditionalSpawnData, IModProxy.ISignHolder {
 
-	private static final DataParameter<ItemStack> HEAD = EntityDataManager.createKey(EntityMelonGolem.class, DataSerializers.ITEMSTACK);
+	private static final EntityDataAccessor<ItemStack> HEAD = SynchedEntityData.defineId(EntityMelonGolem.class, EntityDataSerializers.ITEMSTACK);
 	//	private static final ResourceLocation LOOT = LootTables.register(new ResourceLocation(MelonMod.MODID, "melongolem"));
-	private static final List<DataParameter<ITextComponent>> SIGN_TEXT = Lists.newArrayList(
+	private static final List<EntityDataAccessor<Component>> SIGN_TEXT = Lists.newArrayList(
 
-			EntityDataManager.createKey(EntityMelonGolem.class, DataSerializers.TEXT_COMPONENT),
+			SynchedEntityData.defineId(EntityMelonGolem.class, EntityDataSerializers.COMPONENT),
 
-			EntityDataManager.createKey(EntityMelonGolem.class, DataSerializers.TEXT_COMPONENT),
+			SynchedEntityData.defineId(EntityMelonGolem.class, EntityDataSerializers.COMPONENT),
 
-			EntityDataManager.createKey(EntityMelonGolem.class, DataSerializers.TEXT_COMPONENT),
+			SynchedEntityData.defineId(EntityMelonGolem.class, EntityDataSerializers.COMPONENT),
 
-			EntityDataManager.createKey(EntityMelonGolem.class, DataSerializers.TEXT_COMPONENT)
+			SynchedEntityData.defineId(EntityMelonGolem.class, EntityDataSerializers.COMPONENT)
 
 	);
-	private static final DataParameter<Float> PITCH = EntityDataManager.createKey(EntityMelonGolem.class, DataSerializers.FLOAT);
-	public static BlockState SIGN_TILE_BLOCKSTATE = Blocks.OAK_WALL_SIGN.getDefaultState();
-	public static final SignTileEntity te = new SignTileEntity() {
+	private static final EntityDataAccessor<Float> PITCH = SynchedEntityData.defineId(EntityMelonGolem.class, EntityDataSerializers.FLOAT);
+	public static BlockState SIGN_TILE_BLOCKSTATE = Blocks.OAK_WALL_SIGN.defaultBlockState();
+	public static final SignBlockEntity te = new SignBlockEntity() {
 		@Nonnull
 		@Override
 		public BlockState getBlockState() {
@@ -90,74 +90,74 @@ public class EntityMelonGolem extends GolemEntity implements IRangedAttackMob, I
 		}
 	};
 
-	public EntityMelonGolem(World worldIn) {
+	public EntityMelonGolem(Level worldIn) {
 		this(Objects.requireNonNull(MelonMod.entityTypeMelonGolem), worldIn);
 	}
 
-	protected EntityMelonGolem(EntityType<? extends GolemEntity> p_i48569_1_, World p_i48569_2_) {
-		super(p_i48569_1_, p_i48569_2_);
+	protected EntityMelonGolem(EntityType<? extends AbstractGolem> entity, Level level) {
+		super(entity, level);
 	}
 
 	/*
 	 * underscore required due to a bad reobf catch causing calls to this method to be obfuscated while the method remains the same
 	 */
-	public static AttributeModifierMap.MutableAttribute _registerAttributes() {
-		return MobEntity.func_233666_p_().
-				createMutableAttribute(Attributes.MAX_HEALTH, MelonMod.config.health.get().floatValue()).
-				createMutableAttribute(Attributes.MOVEMENT_SPEED, 0.2F);
+	public static AttributeSupplier.Builder _registerAttributes() {
+		return Mob.createMobAttributes().
+				add(Attributes.MAX_HEALTH, MelonMod.config.health.get().floatValue()).
+				add(Attributes.MOVEMENT_SPEED, 0.2F);
 	}
 
 	@Override
-	protected void registerData() {
-		super.registerData();
-		dataManager.register(HEAD, ItemStack.EMPTY);
-		for (DataParameter<ITextComponent> sign : SIGN_TEXT)
-			dataManager.register(sign, new StringTextComponent(""));
-		dataManager.register(PITCH, rand.nextFloat() * 3.0F);
+	protected void defineSynchedData() {
+		super.defineSynchedData();
+		entityData.define(HEAD, ItemStack.EMPTY);
+		for (EntityDataAccessor<Component> sign : SIGN_TEXT)
+			entityData.define(sign, new TextComponent(""));
+		entityData.define(PITCH, random.nextFloat() * 3.0F);
 	}
 
 	@Override
-	public void setSignText(int index, ITextComponent text) {
-		dataManager.set(SIGN_TEXT.get(index), text);
+	public void setSignText(int index, Component text) {
+		entityData.set(SIGN_TEXT.get(index), text);
 	}
 
 	@Override
 	public int networkID() {
-		return getEntityId();
+		return getId();
 	}
 
 	@Override
-	public ITextComponent getSignText(int index) {
-		return dataManager.get(SIGN_TEXT.get(index));
+	public Component getSignText(int index) {
+		return entityData.get(SIGN_TEXT.get(index));
 	}
 
 	@Override
 	protected void registerGoals() {
 		this.goalSelector.addGoal(1, new RangedAttackGoal(this, 1.25D, 20, 10.0F));
 		this.goalSelector.addGoal(2, new EntityAISearchAndEatMelons(this));
-		this.goalSelector.addGoal(3, new WaterAvoidingRandomWalkingGoal(this, 1.0D, 1.0000001E-5F));
-		this.goalSelector.addGoal(4, new LookAtGoal(this, PlayerEntity.class, 6.0F));
-		this.goalSelector.addGoal(5, new LookRandomlyGoal(this));
+		this.goalSelector.addGoal(3, new WaterAvoidingRandomStrollGoal(this, 1.0D, 1.0000001E-5F));
+		this.goalSelector.addGoal(4, new LookAtPlayerGoal(this, Player.class, 6.0F));
+		this.goalSelector.addGoal(5, new RandomLookAroundGoal(this));
 
-		this.targetSelector.addGoal(1, new NearestAttackableTargetGoal<>(this, MobEntity.class, 10, true, false, (e) -> e instanceof IMob));
+		this.targetSelector.addGoal(1, new NearestAttackableTargetGoal<>(this, Mob.class, 10, true, false, (e) -> e instanceof Enemy));
 	}
 
 	@Override
-	public void attackEntityWithRangedAttack(@Nonnull LivingEntity target, float distanceFactor) {
-		EntityMelonSlice slice = new EntityMelonSlice(this.world, this);
-		double d0 = target.getPosY() + (double) target.getEyeHeight() - 1.100000023841858D;
-		double d1 = target.getPosX() - this.getPosX();
-		double d2 = d0 - slice.getPosY();
-		double d3 = target.getPosZ() - this.getPosZ();
-		float f = MathHelper.sqrt(d1 * d1 + d3 * d3) * 0.2F;
-		slice.shoot(d1, d2 + (double) f, d3, 1.6F, 12.0F);
-		this.playSound(SoundEvents.ENTITY_SNOW_GOLEM_SHOOT, 1.0F, 1.0F / (this.getRNG().nextFloat() * 0.4F + 0.8F));
-		slice.setPositionAndUpdate(slice.getPosX(), slice.getPosY(), slice.getPosZ());
-		MelonMod.spawnNonLivingEntity(world, slice);
+	public void performRangedAttack(@Nonnull LivingEntity target, float distanceFactor) {
+		EntityMelonSlice slice = new EntityMelonSlice(this.level, this);
+		double d0 = target.getY() + (double) target.getEyeHeight() - 1.100000023841858D;
+		double d1 = target.getX() - this.getX();
+		double d2 = d0 - slice.getY();
+		double d3 = target.getZ() - this.getZ();
+		double f = Math.sqrt(d1 * d1 + d3 * d3) * 0.2F;
+		slice.shoot(d1, d2 + f, d3, 1.6F, 12.0F);
+		this.playSound(SoundEvents.SNOW_GOLEM_SHOOT, 1.0F, 1.0F / (this.getRandom().nextFloat() * 0.4F + 0.8F));
+		slice.teleportTo(slice.getX(), slice.getY(), slice.getZ());
+		MelonMod.spawnNonLivingEntity(level, slice);
 	}
 
 	@Override
-	public boolean isShearable(@Nonnull ItemStack item, World world, BlockPos pos) {
+	public boolean isShearable(@Nonnull ItemStack item, Level world, BlockPos vertex) {
 		return !getHead().isEmpty();
 	}
 
@@ -170,23 +170,23 @@ public class EntityMelonGolem extends GolemEntity implements IRangedAttackMob, I
 	@Nullable
 	@Override
 	protected SoundEvent getHurtSound(DamageSource p_184601_1_) {
-		return SoundEvents.ENTITY_SLIME_HURT;
+		return SoundEvents.SLIME_HURT;
 	}
 
 	@Nullable
 	@Override
 	protected SoundEvent getDeathSound() {
-		return SoundEvents.ENTITY_SLIME_DEATH;
+		return SoundEvents.SLIME_DEATH;
 	}
 
 	@Override
-	public int getTalkInterval() {
-		return MelonMod.SIGNS.contains(getHead().getItem()) ? 200 : super.getTalkInterval();
+	public int getAmbientSoundInterval() {
+		return MelonMod.SIGNS.contains(getHead().getItem()) ? 200 : super.getAmbientSoundInterval();
 	}
 
 	@Override
-	public float getSoundPitch() { // protected -> public
-		return super.getSoundPitch();
+	public float getVoicePitch() { // protected -> public
+		return super.getVoicePitch();
 	}
 
 	@Override
@@ -196,139 +196,139 @@ public class EntityMelonGolem extends GolemEntity implements IRangedAttackMob, I
 
 	@Override
 	public void playAmbientSound() {
-		if (world != null && !world.isRemote)
+		if (level != null && !level.isClientSide)
 			MelonMod.network.send(PacketDistributor.TRACKING_ENTITY.with(() -> this), new ClientPacketHandlerMelonAmbientSound(this));
 	}
 
 	@Nonnull
 	@Override
-	public ActionResultType applyPlayerInteraction(PlayerEntity player, Vector3d vec, Hand hand) {
-		if (!MelonMod.config.hats.get() || player.getHeldItemMainhand().getItem() instanceof ShearsItem || player.getHeldItemOffhand().getItem() instanceof ShearsItem)
-			return ActionResultType.FAIL;
-		ItemStack stack = player.getHeldItem(hand);
+	public InteractionResult interactAt(Player player, Vec3 vec, InteractionHand hand) {
+		if (!MelonMod.config.hats.get() || player.getMainHandItem().getItem() instanceof ShearsItem || player.getOffhandItem().getItem() instanceof ShearsItem)
+			return InteractionResult.FAIL;
+		ItemStack stack = player.getItemInHand(hand);
 		if (!stack.isEmpty() && getHead().isEmpty()) {
-			if (Block.getBlockFromItem(stack.getItem()) != Blocks.AIR || MelonMod.SIGNS.contains(stack.getItem())) {
+			if (Block.byItem(stack.getItem()) != Blocks.AIR || MelonMod.SIGNS.contains(stack.getItem())) {
 				setHead(stack);
 				if (!player.isCreative())
-					player.getHeldItem(hand).shrink(1);
-				return ActionResultType.SUCCESS;
+					player.getItemInHand(hand).shrink(1);
+				return InteractionResult.SUCCESS;
 			}
 		} else if (!getHead().isEmpty() && MelonMod.SIGNS.contains(getHead().getItem())) {
 			MelonMod.proxy.openSignHolderGui(this);
-			return ActionResultType.SUCCESS;
+			return InteractionResult.SUCCESS;
 		}
-		return ActionResultType.FAIL;
+		return InteractionResult.FAIL;
 	}
 
 	@Override
-	protected float getStandingEyeHeight(Pose p_213348_1_, EntitySize p_213348_2_) {
+	protected float getStandingEyeHeight(Pose p_213348_1_, EntityDimensions p_213348_2_) {
 		return 1.7F;
 	}
 
 	@Override
 	public ItemStack getHead() {
-		return dataManager.get(HEAD);
+		return entityData.get(HEAD);
 	}
 
 	public void setHead(ItemStack stack) {
 		for (int i = 0; i < 4; ++i)
-			setSignText(i, new StringTextComponent(""));
+			setSignText(i, new TextComponent(""));
 		ItemStack newstack = stack.copy();
 		newstack.setCount(1);
-		dataManager.set(HEAD, newstack);
+		entityData.set(HEAD, newstack);
 	}
 
 	@Nonnull
 	@Override
-	public List<ItemStack> onSheared(@Nullable PlayerEntity player, @Nonnull ItemStack item, World world, BlockPos pos, int fortune) {
+	public List<ItemStack> onSheared(@Nullable Player player, @Nonnull ItemStack item, Level world, BlockPos vertex, int fortune) {
 		List<ItemStack> list = Lists.newArrayList(MelonMod.config.shear.get() ? getHead() : ItemStack.EMPTY);
 		setHead(ItemStack.EMPTY);
 		return list;
 	}
 
 	@Override
-	public void onDeath(@Nonnull DamageSource cause) {
-		super.onDeath(cause);
+	public void die(@Nonnull DamageSource cause) {
+		super.die(cause);
 		ItemStack stack = getHead();
-		if (!world.isRemote && !stack.isEmpty()) {
-			ItemEntity e = new ItemEntity(world, getPosX(), getPosY(), getPosZ(), stack);
-			e.setMotion(e.getMotion().add(
+		if (!level.isClientSide && !stack.isEmpty()) {
+			ItemEntity e = new ItemEntity(level, getX(), getY(), getZ(), stack);
+			e.setDeltaMovement(e.getDeltaMovement().add(
 
-					rand.nextFloat() * 0.05F,
+					random.nextFloat() * 0.05F,
 
-					(rand.nextFloat() - rand.nextFloat()) * 0.1F,
+					(random.nextFloat() - random.nextFloat()) * 0.1F,
 
-					(rand.nextFloat() - rand.nextFloat()) * 0.1F
+					(random.nextFloat() - random.nextFloat()) * 0.1F
 
 			));
-			world.addEntity(e);
+			level.addFreshEntity(e);
 		}
 	}
 
 	@Nonnull
 	@Override
-	public CompoundNBT writeWithoutTypeId(CompoundNBT compound) {
+	public CompoundTag saveWithoutId(CompoundTag compound) {
 		compound.put("head", getHead().serializeNBT());
 		for (int i = 0; i < 4; ++i) {
-			String s = ITextComponent.Serializer.toJson(getSignText(i));
+			String s = Component.Serializer.toJson(getSignText(i));
 			compound.putString("Text" + (i + 1), s);
 		}
-		return super.writeWithoutTypeId(compound);
+		return super.saveWithoutId(compound);
 	}
 
 	@Override
-	public void read(CompoundNBT compound) {
-		setHead(ItemStack.read(compound.getCompound("head")));
+	public void readAdditionalSaveData(CompoundTag compound) {
+		setHead(ItemStack.of(compound.getCompound("head")));
 		for (int i = 0; i < 4; ++i) {
 			String s = compound.getString("Text" + (i + 1));
-			ITextComponent itextcomponent = ITextComponent.Serializer.func_240643_a_(s);
+			Component itextcomponent = Component.Serializer.fromJson(s);
 
 			try {
-				setSignText(i, itextcomponent == null ? new StringTextComponent("") : TextComponentUtils.func_240645_a_(getCommandSource(), itextcomponent, null, 0));
-			} catch (CommandException | CommandSyntaxException var7) {
+				setSignText(i, itextcomponent == null ? new TextComponent("") : ComponentUtils.updateForEntity(createCommandSourceStack(), itextcomponent, null, 0));
+			} catch (CommandRuntimeException | CommandSyntaxException var7) {
 				setSignText(i, itextcomponent);
 			}
 		}
-		super.read(compound);
+		super.readAdditionalSaveData(compound);
 	}
 
 	@Override
-	public void writeSpawnData(PacketBuffer buffer) {
-		buffer.writeItemStack(getHead());
+	public void writeSpawnData(FriendlyByteBuf buffer) {
+		buffer.writeItem(getHead());
 	}
 
 	@Override
-	public void readSpawnData(PacketBuffer additionalData) {
-		setHead(additionalData.readItemStack());
+	public void readSpawnData(FriendlyByteBuf additionalData) {
+		setHead(additionalData.readItem());
 	}
 
 	public float getPitch() {
-		return dataManager.get(PITCH);
+		return entityData.get(PITCH);
 	}
 
 	static class EntityAISearchAndEatMelons extends Goal {
 
-		private final MobEntity parent;
+		private final Mob parent;
 		private final Item melon;
 		private final Block melonblock;
 		private int cooldown;
-		private BlockPos.Mutable pos = new BlockPos.Mutable();
+		private BlockPos.MutableBlockPos vertex = new BlockPos.MutableBlockPos();
 		private boolean foundMelon = false;
 
-		EntityAISearchAndEatMelons(MobEntity entity) {
+		EntityAISearchAndEatMelons(Mob entity) {
 			parent = entity;
-			setMutexFlags(EnumSet.of(Flag.MOVE, Flag.LOOK));
+			setFlags(EnumSet.of(Flag.MOVE, Flag.LOOK));
 			melon = entity instanceof EntityGlisteringMelonGolem ? Items.GLISTERING_MELON_SLICE : Items.MELON_SLICE;
 			melonblock = entity instanceof EntityGlisteringMelonGolem ? MelonMod.glisteringMelonBlock : Blocks.MELON;
 		}
 
 		@Override
-		public boolean shouldExecute() {
+		public boolean canUse() {
 			return MelonMod.config.eats.get() && parent.getHealth() < parent.getMaxHealth();
 		}
 
 		@Override
-		public void resetTask() {
+		public void stop() {
 			cooldown = 0;
 		}
 
@@ -351,35 +351,35 @@ public class EntityMelonGolem extends GolemEntity implements IRangedAttackMob, I
 			if (cooldown > 0)
 				cooldown--;
 			final int radius = 25;
-			AxisAlignedBB area = new AxisAlignedBB(parent.getPosX() - radius, parent.getPosY() - radius, parent.getPosZ() - radius, parent.getPosX() + radius, parent.getPosY() + radius, parent.getPosZ() + radius);
-			List<ItemEntity> items = parent.world.getEntitiesWithinAABB(ItemEntity.class, area);
+			AABB area = new AABB(parent.getX() - radius, parent.getY() - radius, parent.getZ() - radius, parent.getX() + radius, parent.getY() + radius, parent.getZ() + radius);
+			List<ItemEntity> items = parent.level.getEntitiesOfClass(ItemEntity.class, area);
 			for (ItemEntity item : items) {
-				if (parent.getNavigator().noPath() && isMelon(item)) {
-					parent.getNavigator().tryMoveToEntityLiving(item, 1.25F);
-					parent.getLookController().setLookPositionWithEntity(item, 30.0F, 30.0F);
+				if (parent.getNavigation().isDone() && isMelon(item)) {
+					parent.getNavigation().moveTo(item, 1.25F);
+					parent.getLookControl().setLookAt(item, 30.0F, 30.0F);
 				}
-				if (cooldown <= 0 && item.isAlive() && isMelon(item) && item.getBoundingBox().intersects(parent.getBoundingBox().grow(1))) {
+				if (cooldown <= 0 && item.isAlive() && isMelon(item) && item.getBoundingBox().intersects(parent.getBoundingBox().inflate(1))) {
 					boolean flag = item.getItem().getItem() == melonblock.asItem();
 					item.getItem().shrink(1);
-					parent.playSound(SoundEvents.ENTITY_PLAYER_BURP, 1F, 1F);
+					parent.playSound(SoundEvents.PLAYER_BURP, 1F, 1F);
 					parent.heal(MelonMod.config.heal.get().floatValue() * (flag ? 9 : 1));
-					cooldown = 30 + parent.getRNG().nextInt(40);
+					cooldown = 30 + parent.getRandom().nextInt(40);
 				}
 			}
-			if (parent.getNavigator().noPath()) {
+			if (parent.getNavigation().isDone()) {
 				foundMelon = false;
 				search:
 				for (int x = -radius; x < radius; x++)
 					for (int y = -radius; y < radius; y++)
 						for (int z = -radius; z < radius; z++) {
-							pos.setPos(parent.getPosition().add(x, y, z));
-							if (parent.world.isBlockLoaded(pos)) {
-								TileEntity te = parent.world.getTileEntity(pos);
+							vertex.set(parent.blockPosition().offset(x, y, z));
+							if (parent.level.hasChunkAt(vertex)) {
+								BlockEntity te = parent.level.getBlockEntity(vertex);
 								if (te != null) {
 									te.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY, Direction.UP).ifPresent(cap -> {
 										for (int i = 0; i < cap.getSlots(); i++) {
 											if (isMelon(cap.getStackInSlot(i))) {
-												foundMelon = parent.getDistanceSq(pos.getX(), pos.getY(), pos.getZ()) < 4 || parent.getNavigator().tryMoveToXYZ(pos.getX(), pos.getY(), pos.getZ(), 1.25F);
+												foundMelon = parent.distanceToSqr(vertex.getX(), vertex.getY(), vertex.getZ()) < 4 || parent.getNavigation().moveTo(vertex.getX(), vertex.getY(), vertex.getZ(), 1.25F);
 												break;
 											}
 										}
@@ -392,14 +392,14 @@ public class EntityMelonGolem extends GolemEntity implements IRangedAttackMob, I
 			}
 			if (foundMelon) {
 				// Validate
-				if (!parent.world.isBlockLoaded(pos)) {
-					parent.getNavigator().clearPath();
+				if (!parent.level.hasChunkAt(vertex)) {
+					parent.getNavigation().stop();
 					foundMelon = false;
 					return;
 				}
-				TileEntity te = parent.world.getTileEntity(pos);
+				BlockEntity te = parent.level.getBlockEntity(vertex);
 				if (te == null || !te.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY, Direction.UP).isPresent()) {
-					parent.getNavigator().clearPath();
+					parent.getNavigation().stop();
 					foundMelon = false;
 					return;
 				}
@@ -415,21 +415,21 @@ public class EntityMelonGolem extends GolemEntity implements IRangedAttackMob, I
 							}
 						}
 						if (!valid) {
-							parent.getNavigator().clearPath();
+							parent.getNavigation().stop();
 							foundMelon = false;
 							return;
 						}
 
-						if (cooldown <= 0 && parent.getDistanceSq(pos.getX(), pos.getY(), pos.getZ()) < 4) {
+						if (cooldown <= 0 && parent.distanceToSqr(vertex.getX(), vertex.getY(), vertex.getZ()) < 4) {
 							boolean flag = cap.getStackInSlot(i).getItem() == melonblock.asItem();
 							cap.getStackInSlot(i).shrink(1);
-							parent.playSound(SoundEvents.ENTITY_PLAYER_BURP, 1F, 1F);
+							parent.playSound(SoundEvents.PLAYER_BURP, 1F, 1F);
 							parent.heal(MelonMod.config.heal.get().floatValue() * (flag ? 9 : 1));
-							cooldown = 10 + parent.getRNG().nextInt(40);
+							cooldown = 10 + parent.getRandom().nextInt(40);
 						}
 					});
 				} else {
-					parent.getNavigator().clearPath();
+					parent.getNavigation().stop();
 					foundMelon = false;
 				}
 			}
