@@ -1,9 +1,7 @@
 package tamaized.melongolem.common.capability;
 
-import net.minecraft.util.RegistryKey;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.Level;
 import tamaized.melongolem.common.EntityTinyMelonGolem;
 
 import java.util.UUID;
@@ -12,7 +10,7 @@ public class TinyGolemCapabilityHandler implements ITinyGolemCapability {
 
 	private EntityTinyMelonGolem pet;
 
-	private BlockPos pos;
+	private BlockPos vertex;
 	private ResourceLocation dim;
 	private UUID petID;
 
@@ -27,8 +25,8 @@ public class TinyGolemCapabilityHandler implements ITinyGolemCapability {
 	}
 
 	@Override
-	public void markDirty(BlockPos pos, ResourceLocation dim, UUID petID) {
-		this.pos = pos;
+	public void markDirty(BlockPos vertex, ResourceLocation dim, UUID petID) {
+		this.vertex = vertex;
 		this.dim = dim;
 		this.petID = petID;
 	}
@@ -40,7 +38,7 @@ public class TinyGolemCapabilityHandler implements ITinyGolemCapability {
 
 	@Override
 	public BlockPos getLoadPos() {
-		return pos;
+		return vertex;
 	}
 
 	@Override
@@ -50,10 +48,10 @@ public class TinyGolemCapabilityHandler implements ITinyGolemCapability {
 
 	@Override
 	public boolean load(boolean clear) {
-		boolean flag = pet == null && pos != null && petID != null;
+		boolean flag = pet == null && vertex != null && petID != null;
 		if (clear) {
-			dim = World.OVERWORLD.func_240901_a_();
-			pos = null;
+			dim = Level.OVERWORLD.location();
+			vertex = null;
 			petID = null;
 		}
 		return flag;
