@@ -1,11 +1,10 @@
 package tamaized.melongolem.network.server;
 
-
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.network.PacketBuffer;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.TextFormatting;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
 import tamaized.melongolem.IModProxy;
 import tamaized.melongolem.common.EntityMelonGolem;
 import tamaized.melongolem.network.NetworkMessages;
@@ -21,28 +20,28 @@ public class ServerPacketHandlerMelonSign implements NetworkMessages.IMessage<Se
 	}
 
 	@Override
-	public void handle(PlayerEntity player) {
-		Entity entity = player.world.getEntityByID(id);
-		if (entity instanceof EntityMelonGolem && entity.getDistance(player) <= 6)
+	public void handle(Player player) {
+		Entity entity = player.level.getEntity(id);
+		if (entity instanceof EntityMelonGolem && entity.distanceTo(player) <= 6)
 			for (int i = 0; i < lines.length; ++i) {
-				String text = TextFormatting.getTextWithoutFormattingCodes(lines[i]);
-				((EntityMelonGolem) entity).setSignText(i, new StringTextComponent(text == null ? "" : text));
+				String text = ChatFormatting.stripFormatting(lines[i]);
+				((EntityMelonGolem) entity).setSignText(i, new TextComponent(text == null ? "" : text));
 			}
 	}
 
 	@Override
-	public void toBytes(PacketBuffer packet) {
+	public void toBytes(FriendlyByteBuf packet) {
 		packet.writeInt(id);
 		for (int i = 0; i < 4; ++i)
-			packet.writeString(lines[i]);
+			packet.writeUtf(lines[i]);
 	}
 
 	@Override
-	public ServerPacketHandlerMelonSign fromBytes(PacketBuffer packet) {
+	public ServerPacketHandlerMelonSign fromBytes(FriendlyByteBuf packet) {
 		id = packet.readInt();
 		lines = new String[4];
 		for (int i = 0; i < 4; ++i) {
-			this.lines[i] = packet.readString(Short.MAX_VALUE);
+			this.lines[i] = packet.readUtf(Short.MAX_VALUE);
 		}
 		return this;
 	}
